Refactor Email handler into util
