+++ src/main/resources/client_secret.json
@@ -1,0 +1,1 @@
+{"installed":{"client_id":"531305190379-flecg3a0do120lm0as23fj4v0ubt4om5.apps.googleusercontent.com","project_id":"neural-cirrus-198504","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://accounts.google.com/o/oauth2/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"20eJp79KdEF2LO96JyqBhGzJ","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
